requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
python-decouple==3.8
whitenoise==6.6.0
dj-database-url==2.1.0

# Data Science & ML
pandas==2.1.4
numpy==1.24.3
scikit-learn==1.3.2
xgboost==2.0.2
tensorflow==2.15.0
statsmodels==0.14.0
plotly==5.17.0
seaborn==0.13.0
matplotlib==3.8.2

# Additional utilities
requests==2.31.0
python-dateutil==2.8.2
holidays==0.37
beautifulsoup4==4.12.2
celery==5.3.4
redis==5.0.1

# Development
django-debug-toolbar==4.2.0
black==23.11.0
flake8==6.1.0 